Here is a possible commit message for this commit:
"Added schedules configuration data to the application. This includes various settings and options for scheduling tasks, such as frequencies, days of the week, and cron expressions. The data is stored in an array and can be easily accessed and modified by other parts of the application."

This commit message provides a brief summary of what was changed in this commit, including the addition of schedules configuration data to the application. It also mentions that the data is stored in an array and can be easily accessed and modified by other parts of the application.

Here are some possible improvements to this commit message:

* Add more detail about what specific settings or options were added.
* Mention any changes to existing code or functionality as a result of this new configuration data.
* Consider adding a reference to the issue or ticket that this commit is addressing, if applicable.